requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
cryptography==41.0.7
pillow==10.1.0
pytz==2023.3
requests==2.31.0
python-dateutil==2.8.2